[FREETYPE] Better precision, when rendering with antialiasing.

+++ plugins/freetype/glyph.h
@@ -29,27 +29,27 @@ public:
     unsigned int  Index;
     /*! Bitmap width
      */
-    float  Width;
+    double  Width;
     /*! Bitmap height
      */
-    float  Height;
+    double  Height;
     /*! A width in texture coordinates
      */
-    float TexCoordinateWidth;
+    double TexCoordinateWidth;
     /*! A height in texture coordinates
      */
-    float TexCoordinateHeight;
+    double TexCoordinateHeight;
     /*! A vertical bearing  as
         distance from baseline to top point of glyph
      */ 
-    float BearingY;
+    double BearingY;
     /*! A vertical distance from baseline to bottom point
         of glyph
      */
-    float Descender;
+    double Descender;
     /*! How long should pen move, after glyph was rendered
      */
-    float AdvanceX;
+    double AdvanceX;
 
     /*! Maximum Y of bounding box
      */
